+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.73
+### is.codion.swing.common.ui
+- Dialogs.displayException() removed.
 
 ## 0.18.72
 ### is.codion.common.reactive